==> 4_CRF.ipynb <==
!pip install xgboost

!pip install sklearn-crfsuite

import xgboost as xgb
import pandas as pd
from sklearn.metrics import mean_absolute_error
from sklearn.model_selection import train_test_split
import networkx as nx
import matplotlib.pyplot as plt
from sklearn_crfsuite import CRF

dataset = pd.read_csv("final_merged_phq1_9.csv")
print(dataset.columns)

X = dataset.drop(columns=[f"target0{i}" for i in range(1, 10)]) 
X = X.iloc[:, 2:]
Y = dataset[[f"target0{i}" for i in range(1, 10)]]

X.shape

X_train, X_test, Y_train, Y_test = train_test_split(X, Y, test_size=0.2, random_state=42)

**predict individual question using xgboost**

individual_pred = {}
for target in range(1, 10):
    target = "target0" + str(target)
    
    # Define and train the model
    model = xgb.XGBRegressor(objective="reg:squarederror", random_state=42)
    model.fit(X_train, Y_train[target])
        
    individual_pred[target] = model.predict(X_test)   

# Combine predictions into a DataFrame
individual_preds_df = pd.DataFrame(individual_pred)
print(individual_preds_df.head())

**CRF for structured output**

# Convert each row into a sequence of features
X_sequences = []

for _, row in X.iterrows():
    
    # Create a sequence of 9 feature dictionaries for each instance
    sequence = []
    
    for i in range(1, 10):  # 9 labels (target01 to target09)
        feature_dict = {col: row[col] for col in X.columns}
        sequence.append(feature_dict)
    X_sequences.append(sequence)
    
    # Create Y_sequences (list of sequences of labels)
    Y_sequences = Y.apply(lambda row: [str(int(value)) for value in row], axis=1).tolist()

print(f"Number of feature sequences: {len(X_sequences)}")
print(f"Number of label sequences: {len(Y_sequences)}")

crf = CRF(
        algorithm='lbfgs',  # Optimization algorithm
        max_iterations=100,  # Maximum number of iterations
        all_possible_transitions=True  # Allow transitions between all labels
    )

# Fit the CRF model
crf.fit(X_sequences, Y_sequences)

# Predict on new data
structured_pred = crf.predict(X_sequences)

**Accuracy Comparison**

# Evaluate Individual Predictions
individual_mae = {}

for target in range(1, 10):
    
    target = f"target0{target}"
    mae = mean_absolute_error(Y_test[target], individual_preds_df[target])
    individual_mae[target] = mae

average_mae = sum(individual_mae.values()) / len(individual_mae)
print(f"Average MAE across individual predictions: {average_mae:.4f}")

# Convert CRF predictions to DataFrame
structured_preds_df = pd.DataFrame(structured_pred, columns=[f"target0{i}" for i in range(1, 10)])

# Convert predictions and ground truth back to numerical format
structured_preds_df = structured_preds_df.apply(pd.to_numeric)
Y_numeric = Y.apply(pd.to_numeric)

# Evaluate Structured Predictions
structured_mae = mean_absolute_error(Y_numeric, structured_preds_df)

print("Structured Prediction MAE:", structured_mae)
